Moving UTF-16 SymbolTable extensions directly onto base class.
